requirements.txt
pandas>=1.3.0
pytest>=7.0.0
python-dotenv>=1.1.0 